+++ src/pages/MainPage/LankingList/EplRanking.css
@@ -1,25 +1,3 @@
-.list {
-  width: 20%;
-  height: 100vh;
-
-  margin-left: 2%;
-}
-.Topleagues {
-  font-size: 2.3rem;
-  font-weight: 600;
-}
-
-.Epl,
-.Laliga,
-.SerieA {
-  font-size: 1.6rem;
-  font-weight: 400;
-
-  cursor: pointer;
-  margin-bottom: 10%;
-}
-
-/* 우측 랭킹 리스트 */
 .Ranking {
   width: 20%;
