+++ src/pages/MainPage/LeagueInformation/Epl.css
@@ -42,3 +42,7 @@
 
   margin-right: 15px;
 }
+
+Link {
+  text-decoration: none;
+}
