+++ src/pages/MainPage/LeagueInformation/Laliga.css
@@ -1,0 +1,3 @@
+.Laliga_body {
+  display: flex;
+}
